Fix from review
